Provide http client from core
