+++ workshops/AndroidKotlinWorkshop/app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".ContributorsActivity">
+        <activity android:name=".contributors.ContributorsActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
