+++ workshops/AndroidKotlinWorkshop/app/src/main/java/com/novoda/workshop/contributors/ContributorsActivity.kt
@@ -1,17 +1,18 @@
-package com.novoda.workshop
+package com.novoda.workshop.contributors
 
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
+import com.novoda.workshop.GitHubService
+import com.novoda.workshop.R
+import com.novoda.workshop.core.NetworkDependencyProvider
 import kotlinx.android.synthetic.main.activity_contributers.*
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
-import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
-import java.util.*
 import kotlin.coroutines.CoroutineContext
 
 class ContributorsActivity : AppCompatActivity(), CoroutineScope {
@@ -42,20 +43,7 @@ class ContributorsActivity : AppCompatActivity(), CoroutineScope {
     private fun createService(): GitHubService {
         val userName = "XXX"
         val token = "XXX"
-        val authToken =
-            "Basic " + Base64.getEncoder().encode("$userName:$token".toByteArray()).toString(
-                Charsets.UTF_8
-            )
-        val httpClient = OkHttpClient.Builder()
-            .addInterceptor { chain ->
-                val original = chain.request()
-                val builder = original.newBuilder()
-                    .header("Accept", "application/vnd.github.v3+json")
-                    .header("Authorization", authToken)
-                val request = builder.build()
-                chain.proceed(request)
-            }.build()
-
+        val httpClient = NetworkDependencyProvider(userName, token).provideHttpClient()
 
         val retrofit = Retrofit.Builder()
             .baseUrl("https://api.github.com")
